requirements.txt
fastapi
uvicorn
openai
pydantic
transformers
tensorflow
google-generativeai
faiss-cpu
sentence-transformers
